+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Fix sound mixing for uncontrolled time
 * Fix concurrent access to action queue causing a crash in libTAS
 * Vulkan functions loaded through dlsym are correctly hooked
+* Protect pthread keys access
 
 ## [1.4.8] - 2026-05-14
 ### Added